fix: address Greptile routing feedback
