+++ react-router.config.ts
@@ -1,10 +1,11 @@
 import type { Config } from "@react-router/dev/config";
 import { pages } from "./src/seo/pageMetaMap.generated";
+import { isIndexablePagePath } from "./src/seo/meta";
 
 const prerenderPaths: string[] = [];
 for (const page of pages) {
-  const pathSegments = page.pathWithoutLang.split("/").filter(Boolean);
-  if (pathSegments.length <= 1) {
+  // Practice and exam routes are intentionally noindex and stay on the SPA fallback.
+  if (isIndexablePagePath(page.pathWithoutLang)) {
     prerenderPaths.push(new URL(page.canonicalUrl).pathname);
   }
 }
