+++ src/routes/legacy-exam-redirect.tsx
@@ -1,0 +1,1 @@
+export { default } from "./language-redirect";
